+++ src/main/java/net/dungeonz/access/InGameHudAccess.java
@@ -1,0 +1,11 @@
+package net.dungeonz.access;
+
+public interface InGameHudAccess {
+
+    public void setDungeonCountdownTicks(int dungeonCountdown);
+
+    public int getDungeonCountdownTicks();
+
+    public int getDungeonCountdownRemainingTicks();
+
+}