+++ src/main/java/net/dungeonz/config/DungeonzConfig.java
@@ -9,7 +9,12 @@ import me.shedaniel.cloth.clothconfig.shadowed.blue.endless.jankson.Comment;
 public class DungeonzConfig implements ConfigData {
 
     public boolean defaultDungeons = true;
-    @Comment("Crops won't grow")
+    @Comment("In ticks, used for preloading")
+    public int defaultDungeonTeleportCountdown = 200;
+    public int countdownX = 0;
+    public int countdownY = 0;
+    public float countdownSize = 6.0f;
+    @Comment("Among other things crops won't grow")
     public boolean devMode = false;
 
 }