+++ src/main/java/net/dungeonz/init/RenderInit.java
@@ -5,9 +5,11 @@ import net.dungeonz.block.render.DungeonPortalRenderer;
 import net.dungeonz.block.render.DungeonSpawnerRenderer;
 import net.dungeonz.block.screen.DungeonPortalScreen;
 import net.dungeonz.item.DungeonCompassItem;
+import net.dungeonz.util.RenderHelper;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
+import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 import net.minecraft.client.gui.screen.ingame.HandledScreens;
 import net.minecraft.client.item.CompassAnglePredicateProvider;
 import net.minecraft.client.item.ModelPredicateProviderRegistry;
@@ -31,6 +33,9 @@ public class RenderInit {
         ModelPredicateProviderRegistry.register(ItemInit.DUNGEON_COMPASS, new Identifier("angle"), new CompassAnglePredicateProvider((world, stack, entity) -> {
             return DungeonCompassItem.createGlobalDungeonStructurePos(world, stack);
         }));
+        HudRenderCallback.EVENT.register((drawContext, tickDelta) -> {
+            RenderHelper.renderDungeonCountdown(drawContext, tickDelta);
+        });
     }
 
 }